# openapi.yaml
openapi: 3.0.0
info:
  title: Digital Marketplace API - Test Spec
  version: '1.0'
  description: OpenAPI spec for testing all implemented endpoints (URL / input / output).
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/content/my-library:
    get:
      summary: List licenses/products owned by user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of licenses
  /api/content/licenses/{licenseId}/revoke:
    post:
      summary: Revoke a license (admin only)
      parameters:
        - in: path
          name: licenseId
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Revoked
  /api/content/signed/{productId}:
    get:
      summary: Get a signed streaming URL for a product (owner only)
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signed URL
  /api/content/stream/{token}:
    get:
      summary: Stream content by token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Binary stream
...